refactor: ディレクトリの位置の変更
- viewとlogicディレクトリを_components下に移動
