+++ src/app/search/[keyword]/_components/logic/Logic.tsx
