+++ src/app/search/[keyword]/_components/view/index.tsx
