+++ src/app/search/[keyword]/page.tsx
@@ -1,4 +1,4 @@
-import SearchPostKeywordLogic from "./logic/Logic";
+import SearchPostKeywordLogic from "./_components/logic/Logic";
 
 export default function Page({ params }: { params: Promise<{ keyword: string }> }) {
   return <SearchPostKeywordLogic params = {params}/>;
