+++ src/app/search/_components/view/index.tsx
